+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 1. Method to delete Company `companies.delete({ id })`
 2. Method to attach Contact to Company `companies.attachContact({ companyId, contactId })`
 3. Method to detach Contact from Company `companies.detachContact({ companyId, contactId })`
+4. Method to list attached Contacts for Company `companies.listAttachedContacts({ companyId, page, perPage })`
 
 ## Changed